requirements.txt
# Voice Assistant Requirements

# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
langchain>=0.3.0
langchain-community>=0.3.0

# Audio Processing
whisper-openai>=20231117
sounddevice>=0.5.0
pyaudio>=0.2.14
speechrecognition>=3.14.0

# Text Processing
nltk>=3.8.0
rich>=13.0.0

# TTS
suno-bark>=0.0.1

# HTTP & API
requests>=2.31.0
httpx>=0.27.0

# Configuration
pyyaml>=6.0

# Utilities
numpy>=1.24.0
queue

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Audio Processing for Web
soundfile>=0.12.0

# Optional GPU support
# nvidia-ml-py3  # For NVIDIA GPU monitoring
